refactor(platform): share connector account mutation mapping
