+++ turbo/apps/platform/src/views/okou-page/components/settings/add-connection-dialog.tsx
@@ -65,6 +65,7 @@ import { detach, onDomEventFn, Reason } from "../../../../signals/utils.ts";
 import { ConnectorHelpText } from "./connector-help-text.tsx";
 import { i18n } from "../../../../i18n/index.ts";
 import {
+  connectorAccountMutationFor,
   connectorAccountLabel$,
   setConnectorAccountLabel$,
   type ConnectorAccountConnectMode,
@@ -187,20 +188,6 @@ type ConnectModalContentProps = {
   accountMode?: ConnectorAccountConnectMode;
   accountLabel: string;
 };
-
-function connectorAccountMutationFor(
-  mode: ConnectorAccountConnectMode | undefined,
-  accountLabel: string,
-): ConnectorAccountMutationIntent | undefined {
-  if (!mode) {
-    return undefined;
-  }
-  if (mode.kind === "reconnect") {
-    return { intent: "reconnect", connectionId: mode.account.id };
-  }
-  const displayName = accountLabel.trim();
-  return displayName ? { intent: "add", displayName } : { intent: "add" };
-}
 
 function accountLabelRequired(
   mode: ConnectorAccountConnectMode | undefined,
